web_interface/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
Pillow>=10.0.0
ultralytics>=8.0.0
google-cloud-vision>=3.0.0
requests>=2.31.0
numpy>=1.24.0
opencv-python>=4.8.0
